updates for better control of vis features
Former-commit-id: 50e92c0f00a8b79b702cc2cb465eb39fdb89dbad

+++ src/ves/xplorer/event/viz/cfdContourBase.cxx
@@ -250,7 +250,7 @@ void cfdContourBase::UpdateCommand()
     activeModelDVP = objectCommand->GetDataValuePair( "Position" );
     double planePosition;
     activeModelDVP->GetData( planePosition );
-    SetRequestedValue( static_cast< int >( planePosition ) );
+    SetRequestedValue( planePosition );
 
     activeModelDVP = objectCommand->GetDataValuePair( "Plane Option" );
     if( activeModelDVP )
