+++ src/ves/xplorer/event/viz/cfdVectorBase.cxx
@@ -121,7 +121,7 @@ void cfdVectorBase::UpdateCommand()
     activeModelDVP = objectCommand->GetDataValuePair( "Position" );
     double planePosition;
     activeModelDVP->GetData( planePosition );
-    SetRequestedValue( static_cast< int >( planePosition ) );
+    SetRequestedValue( planePosition );
 
     activeModelDVP = objectCommand->GetDataValuePair( "Plane Option" );
     if( activeModelDVP )
